+++ src/views/user/SignupPage.vue
@@ -38,7 +38,7 @@
         required
       ></v-text-field>
 
-      <CustomDatePicker
+      <CustomDatePickerComp
         @pickup-date="setBirthYmd"
         :p-label-name="pPickerLabelName"
         :p-init-value="pInitValue"
@@ -80,14 +80,14 @@
 
 <script>
 import userApi from "@/api/UserAPI";
-import CustomDatePicker from "@/components/common/CustomDatePicker.vue";
+import CustomDatePickerComp from "@/components/common/CustomDatePickerComp.vue";
 import router from "@/router";
 
 // id중복체크 - 자동으로 체크하기 https://pozafly.github.io/tripllo/(6)login3-vue/
 // 참고자료 : https://vuetifyjs.com/en/components/forms/#vuelidate
 export default {
   components: {
-    CustomDatePicker,
+    CustomDatePickerComp,
   },
   data: () => {
     return {
